Updated after offline review
