+++ bots/pr/src/main/java/org/openjdk/skara/bots/pr/IntegrateCommand.java
@@ -78,8 +78,9 @@ public class IntegrateCommand implements CommandHandler {
             var sanitizedUrl = URLEncoder.encode(pr.repository().webUrl().toString(), StandardCharsets.UTF_8);
             var path = scratchPath.resolve("pr.integrate").resolve(sanitizedUrl);
 
+            var seedPath = bot.seedStorage().orElse(scratchPath.resolve("seeds"));
             var prInstance = new PullRequestInstance(path,
-                                                     bot.seedStorage().orElse(scratchPath.resolve("seeds")),
+                                                     new HostedRepositoryPool(seedPath),
                                                      pr,
                                                      bot.ignoreStaleReviews());
             var localHash = prInstance.commit(censusInstance.namespace(), censusInstance.configuration().census().domain(), null);
