+++ bots/pr/src/main/java/org/openjdk/skara/bots/pr/LabelerWorkItem.java
@@ -22,7 +22,7 @@
  */
 package org.openjdk.skara.bots.pr;
 
-import org.openjdk.skara.forge.PullRequest;
+import org.openjdk.skara.forge.*;
 
 import java.io.*;
 import java.nio.file.Path;
@@ -63,8 +63,9 @@ public class LabelerWorkItem extends PullRequestWorkItem {
             return;
         }
         try {
+            var seedPath = bot.seedStorage().orElse(scratchPath.resolve("seeds"));
             var prInstance = new PullRequestInstance(scratchPath.resolve("labeler"),
-                                                     bot.seedStorage().orElse(scratchPath.resolve("seeds")),
+                                                     new HostedRepositoryPool(seedPath),
                                                      pr,
                                                      bot.ignoreStaleReviews());
             var newLabels = getLabels(prInstance);
